+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+UPDATE 3/13/2022:
+
+- Blocks black list added to config. By default Concrete and Wood blocks (modded blocks, not textures) will not rust.
+
 UPDATE 3/2/2022:
 
 - Rusting of powered grids can be disabled in config.